refactor: simplify platform support in e2e package
